+++ test/flutter_bluetooth_plugin_models_test.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:flutter_bluetooth_plugin/flutter_bluetooth_plugin_models.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -38,5 +40,40 @@ void main() {
     expect(info.isSupported, isTrue);
     expect(info.isBleSupported, isTrue);
     expect(info.isDiscovering, isFalse);
+  });
+
+  test('Bluetooth byte payload fields use Uint8List', () {
+    final scanResult = BluetoothScanResult.fromMap(<String, dynamic>{
+      'device': <String, dynamic>{'id': 'device-1'},
+      'rssi': -55,
+      'manufacturerData': <String, Object>{
+        '76': Uint8List.fromList(<int>[1, 2, 3]),
+      },
+      'serviceData': <String, Object>{
+        '0000fff0-0000-1000-8000-00805f9b34fb': <int>[4, 5],
+      },
+    });
+
+    expect(scanResult.manufacturerData[76], isA<Uint8List>());
+    expect(scanResult.manufacturerData[76], <int>[1, 2, 3]);
+    expect(
+      scanResult.serviceData['0000fff0-0000-1000-8000-00805f9b34fb'],
+      isA<Uint8List>(),
+    );
+    expect(
+      scanResult.serviceData['0000fff0-0000-1000-8000-00805f9b34fb'],
+      <int>[4, 5],
+    );
+
+    final characteristic = BluetoothGattCharacteristic.fromMap(
+      <String, dynamic>{
+        'uuid': '0000fff1-0000-1000-8000-00805f9b34fb',
+        'serviceUuid': '0000fff0-0000-1000-8000-00805f9b34fb',
+        'value': ByteData.sublistView(Uint8List.fromList(<int>[6, 7])),
+      },
+    );
+
+    expect(characteristic.value, isA<Uint8List>());
+    expect(characteristic.value, <int>[6, 7]);
   });
 }